Refactor: Payment files
